Added metric to track blob tree size

+++ module/metrics/noop.go
@@ -149,6 +149,6 @@ func (nc *NoopCollector) DiskSize(uint64)                                       
 func (nc *NoopCollector) ExecutionBlockDataUploadStarted()                                      {}
 func (nc *NoopCollector) ExecutionBlockDataUploadFinished(dur time.Duration)                    {}
 func (nc *NoopCollector) ExecutionDataAddStarted()                                              {}
-func (nc *NoopCollector) ExecutionDataAddFinished(time.Duration, bool, int)                     {}
+func (nc *NoopCollector) ExecutionDataAddFinished(time.Duration, bool, uint64)                  {}
 func (nc *NoopCollector) ExecutionDataGetStarted()                                              {}
-func (nc *NoopCollector) ExecutionDataGetFinished(time.Duration, bool, int)                     {}
+func (nc *NoopCollector) ExecutionDataGetFinished(time.Duration, bool, uint64)                  {}
